17_compute/opencode/pyproject.toml
[project]
name = "opencode-example"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "chalkcompute>=1.1.1",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]
